combine CompactionUtils and InnerSpaceCompactionUtils together
