+++ server/src/test/java/org/apache/iotdb/db/engine/compaction/inner/InnerSeqCompactionTest.java
@@ -22,7 +22,7 @@ package org.apache.iotdb.db.engine.compaction.inner;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.engine.cache.ChunkCache;
 import org.apache.iotdb.db.engine.cache.TimeSeriesMetadataCache;
-import org.apache.iotdb.db.engine.compaction.inner.utils.InnerSpaceCompactionUtils;
+import org.apache.iotdb.db.engine.compaction.CompactionUtils;
 import org.apache.iotdb.db.engine.compaction.performer.ReadChunkCompactionPerformer;
 import org.apache.iotdb.db.engine.compaction.utils.CompactionCheckerUtils;
 import org.apache.iotdb.db.engine.compaction.utils.CompactionClearUtils;
@@ -37,6 +37,7 @@ import org.apache.iotdb.db.metadata.path.PartialPath;
 import org.apache.iotdb.db.service.IoTDB;
 import org.apache.iotdb.db.utils.EnvironmentUtils;
 import org.apache.iotdb.tsfile.common.conf.TSFileDescriptor;
+import org.apache.iotdb.tsfile.exception.write.WriteProcessException;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSEncoding;
 import org.apache.iotdb.tsfile.read.TimeValuePair;
@@ -113,7 +114,7 @@ public class InnerSeqCompactionTest {
   }
 
   @Test
-  public void testDeserializePage() throws MetadataException, IOException {
+  public void testDeserializePage() throws MetadataException, IOException, WriteProcessException {
 
     long chunkSizeLowerBoundInCompaction =
         IoTDBDescriptor.getInstance().getConfig().getChunkSizeLowerBoundInCompaction();
@@ -229,9 +230,9 @@ public class InnerSeqCompactionTest {
                             && timeValuePair.getTimestamp() <= 300L);
               }
               new ReadChunkCompactionPerformer(sourceResources, targetTsFileResource).perform();
-              InnerSpaceCompactionUtils.moveTargetFile(targetTsFileResource, COMPACTION_TEST_SG);
-              InnerSpaceCompactionUtils.combineModsInCompaction(
-                  sourceResources, targetTsFileResource);
+              CompactionUtils.moveTargetFile(
+                  Collections.singletonList(targetTsFileResource), true, COMPACTION_TEST_SG);
+              CompactionUtils.combineModsInInnerCompaction(sourceResources, targetTsFileResource);
               List<TsFileResource> targetTsFileResources = new ArrayList<>();
               targetTsFileResources.add(targetTsFileResource);
               // check data
@@ -349,7 +350,8 @@ public class InnerSeqCompactionTest {
 
   @Test
   public void testAppendPage()
-      throws IOException, MetadataException, InterruptedException, StorageEngineException {
+      throws IOException, MetadataException, InterruptedException, StorageEngineException,
+          WriteProcessException {
 
     for (int toMergeFileNum : toMergeFileNums) {
       for (CompactionTimeseriesType compactionTimeseriesType : compactionTimeseriesTypes) {
@@ -455,9 +457,9 @@ public class InnerSeqCompactionTest {
                       timeValuePair.getTimestamp() >= 250L && timeValuePair.getTimestamp() <= 300L);
             }
             new ReadChunkCompactionPerformer(toMergeResources, targetTsFileResource).perform();
-            InnerSpaceCompactionUtils.moveTargetFile(targetTsFileResource, COMPACTION_TEST_SG);
-            InnerSpaceCompactionUtils.combineModsInCompaction(
-                toMergeResources, targetTsFileResource);
+            CompactionUtils.moveTargetFile(
+                Collections.singletonList(targetTsFileResource), true, COMPACTION_TEST_SG);
+            CompactionUtils.combineModsInInnerCompaction(toMergeResources, targetTsFileResource);
             List<TsFileResource> targetTsFileResources = new ArrayList<>();
             targetTsFileResources.add(targetTsFileResource);
             CompactionCheckerUtils.checkDataAndResource(sourceData, targetTsFileResources);
@@ -612,7 +614,8 @@ public class InnerSeqCompactionTest {
 
   @Test
   public void testAppendChunk()
-      throws IOException, IllegalPathException, MetadataException, StorageEngineException {
+      throws IOException, IllegalPathException, MetadataException, StorageEngineException,
+          WriteProcessException {
     long prevChunkPointNumLowerBoundInCompaction =
         IoTDBDescriptor.getInstance().getConfig().getChunkPointNumLowerBoundInCompaction();
     IoTDBDescriptor.getInstance().getConfig().setChunkPointNumLowerBoundInCompaction(1);
@@ -731,9 +734,9 @@ public class InnerSeqCompactionTest {
                             && timeValuePair.getTimestamp() <= 300L);
               }
               new ReadChunkCompactionPerformer(toMergeResources, targetTsFileResource).perform();
-              InnerSpaceCompactionUtils.moveTargetFile(targetTsFileResource, COMPACTION_TEST_SG);
-              InnerSpaceCompactionUtils.combineModsInCompaction(
-                  toMergeResources, targetTsFileResource);
+              CompactionUtils.moveTargetFile(
+                  Collections.singletonList(targetTsFileResource), true, COMPACTION_TEST_SG);
+              CompactionUtils.combineModsInInnerCompaction(toMergeResources, targetTsFileResource);
               List<TsFileResource> targetTsFileResources = new ArrayList<>();
               targetTsFileResources.add(targetTsFileResource);
               CompactionCheckerUtils.checkDataAndResource(sourceData, targetTsFileResources);
